requirements.txt
streamlit
pandas
openpyxl
requests
openai
python-dotenv